check for duplicate selectiona and measurement and skip if any exist

+++ py/experimentBudgetModified.py
@@ -135,37 +135,44 @@ for index, prog in enumerate(projects_list):
         reduction = repetitions
 
         for run in range(REPEATS):
-            pTime, rTime, sel = fastr.fastPlusPlus(inputFile, dim=dim, B=B)
             sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST++", reduction, run+1)
-            pickle.dump(sel, open(sOut, "wb"))
             tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST++", reduction, run+1)
+            if os.path.exists(sOut) and os.path.exists(tOut):
+                continue
+            pTime, rTime, sel = fastr.fastPlusPlus(inputFile, dim=dim, B=B)
+            pickle.dump(sel, open(sOut, "wb"))
             pickle.dump((pTime, rTime), open(tOut, "wb"))
             print("FAST++", reduction, pTime, rTime)
 
         for run in range(REPEATS):
-            pTime, rTime, sel = fastr.fastCS(inputFile, dim=dim, B=B)
             sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-CS", reduction, run+1)
-            pickle.dump(sel, open(sOut, "wb"))
             tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-CS", reduction, run+1)
+            if os.path.exists(sOut) and os.path.exists(tOut):
+                continue
+            pTime, rTime, sel = fastr.fastCS(inputFile, dim=dim, B=B)
+            pickle.dump(sel, open(sOut, "wb"))
             pickle.dump((pTime, rTime), open(tOut, "wb"))
             print("FAST-CS", reduction, pTime, rTime)
 
 
         for run in range(REPEATS):
-            pTime, rTime, sel = fastr.fast_pw(inputFile, r, b, bbox=True, k=k, memory=True, B=B)
             sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-pw", reduction, run+1)
-            pickle.dump(sel, open(sOut, "wb"))
             tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-pw", reduction, run+1)
+            if os.path.exists(sOut) and os.path.exists(tOut):
+                continue
+            pTime, rTime, sel = fastr.fast_pw(inputFile, r, b, bbox=True, k=k, memory=True, B=B)
+            pickle.dump(sel, open(sOut, "wb"))
             pickle.dump((pTime, rTime), open(tOut, "wb"))
             print("FAST-pw", reduction, pTime, rTime)
 
 
         for run in range(REPEATS):
-            pTime, rTime, sel = fastr.fast_(inputFile, all_, r=r, b=b, bbox=True, k=k, memory=True, B=B)
-            print("FAST-all:", sel)
             sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-all", reduction, run+1)
-            pickle.dump(sel, open(sOut, "wb"))
             tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-all", reduction, run+1)
+            if os.path.exists(sOut) and os.path.exists(tOut):
+                continue
+            pTime, rTime, sel = fastr.fast_(inputFile, all_, r=r, b=b, bbox=True, k=k, memory=True, B=B)
+            pickle.dump(sel, open(sOut, "wb"))
             pickle.dump((pTime, rTime), open(tOut, "wb"))
             print("FAST-all", reduction, pTime, rTime)
 
@@ -216,39 +223,46 @@ for index, prog in enumerate(projects_list):
             reduction = MIN_PERCENTAGE_OF_TEST_PRESERVED
             
             for run in range(REPEATS):
-                pTime, rTime, sel = fastr.fastPlusPlus(inputFile, dim=dim, B=B)
                 sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST++", reduction, run+1)
-                pickle.dump(sel, open(sOut, "wb"))
                 tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST++", reduction, run+1)
+                if os.path.exists(sOut) and os.path.exists(tOut):
+                    continue
+                pTime, rTime, sel = fastr.fastPlusPlus(inputFile, dim=dim, B=B)
+                pickle.dump(sel, open(sOut, "wb"))
                 pickle.dump((pTime, rTime), open(tOut, "wb"))
                 print("FAST++", reduction, pTime, rTime)
 
             for run in range(REPEATS):
-                pTime, rTime, sel = fastr.fastCS(inputFile, dim=dim, B=B)
                 sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-CS", reduction, run+1)
-                pickle.dump(sel, open(sOut, "wb"))
                 tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-CS", reduction, run+1)
+                if os.path.exists(sOut) and os.path.exists(tOut):
+                    continue
+                pTime, rTime, sel = fastr.fastCS(inputFile, dim=dim, B=B)
+                pickle.dump(sel, open(sOut, "wb"))
                 pickle.dump((pTime, rTime), open(tOut, "wb"))
                 print("FAST-CS", reduction, pTime, rTime)
 
 
             for run in range(REPEATS):
-                pTime, rTime, sel = fastr.fast_pw(inputFile, r, b, bbox=True, k=k, memory=True, B=B)
                 sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-pw", reduction, run+1)
-                pickle.dump(sel, open(sOut, "wb"))
                 tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-pw", reduction, run+1)
+                if os.path.exists(sOut) and os.path.exists(tOut):
+                    continue
+                pTime, rTime, sel = fastr.fast_pw(inputFile, r, b, bbox=True, k=k, memory=True, B=B)
+                pickle.dump(sel, open(sOut, "wb"))
                 pickle.dump((pTime, rTime), open(tOut, "wb"))
                 print("FAST-pw", reduction, pTime, rTime)
 
 
             for run in range(REPEATS):
-                pTime, rTime, sel = fastr.fast_(inputFile, all_, r=r, b=b, bbox=True, k=k, memory=True, B=B)
                 sOut = "{}/{}-{}-{}.pickle".format(sPath, "FAST-all", reduction, run+1)
-                pickle.dump(sel, open(sOut, "wb"))
                 tOut = "{}/{}-{}-{}.pickle".format(tPath, "FAST-all", reduction, run+1)
+                if os.path.exists(sOut) and os.path.exists(tOut):
+                    continue
+                pTime, rTime, sel = fastr.fast_(inputFile, all_, r=r, b=b, bbox=True, k=k, memory=True, B=B)
+                pickle.dump(sel, open(sOut, "wb"))
                 pickle.dump((pTime, rTime), open(tOut, "wb"))
                 print("FAST-all", reduction, pTime, rTime)
-
 
 f = open(f"./loose_budget.json", "w")
 f.write(json.dumps(LOOSE_BUDGET))
